tikv_util: use local thread prefix matching
