+++ components/raftstore/src/store/worker/split_controller.rs
@@ -22,7 +22,7 @@ use tikv_util::{
     debug, info,
     metrics::ThreadInfoStatistics,
     store::{QueryStats, is_read_query},
-    thread_name_prefix::matches_thread_name_prefix,
+    sys::thread::matches_thread_name_prefix,
     time::Instant,
     warn,
 };
@@ -38,6 +38,8 @@ use crate::store::{
 
 const DEFAULT_MAX_SAMPLE_LOOP_COUNT: usize = 10000;
 pub const TOP_N: usize = 10;
+const GRPC_SERVER_THREAD: &str = "grpc-server";
+const UNIFIED_READ_POOL_THREAD: &str = "unified-read";
 
 // It will return prefix sum of the given iter,
 // `read` is a function to process the item from the iter.
@@ -773,8 +775,8 @@ impl AutoSplitController {
         // Prepare some diagnostic info.
         thread_stats.record();
         let (grpc_thread_usage, unified_read_pool_thread_usage) = (
-            Self::collect_thread_usage(thread_stats, "grpc-server"),
-            Self::collect_thread_usage(thread_stats, "unified-read-po"),
+            Self::collect_thread_usage(thread_stats, GRPC_SERVER_THREAD),
+            Self::collect_thread_usage(thread_stats, UNIFIED_READ_POOL_THREAD),
         );
         // Update first before calculating the latest average gRPC poll CPU usage.
         self.update_grpc_thread_usage(grpc_thread_usage);
