+++ components/tikv_util/src/lib.rs
@@ -62,7 +62,6 @@ pub mod store;
 pub mod stream;
 pub mod sys;
 pub mod thread_group;
-pub mod thread_name_prefix;
 pub mod time;
 pub mod timer;
 pub mod topn;
